WIP: Implementing Dagger2 configuration
Attempting to use Dagger2 to handle DI. Add @Component elements and the necessary
@Inject annotations and generated class file names. As part of this testing additional test
methods have been added to validate the DI framework is in fact generating the relevant
mock instances of cache and remote facets and that they are being injected into the conductor
instance. Basic structure for the current configuration is a ConductorComponent which will
include various feature modules (at this point only contains the MovieListModule). These
feature Modules will be overriden to provide the final injection values of production instances.

Some take aways from this...
* Code generated for DI on test packaging does not resolve in Android Studio
* At the moment setting  up a simple injection structure seems quite long winded
* @Component, and segmented injection may come more to the fore as the functionality
    inside the application fills out.
* DI specification in the test packages does seem to need to be a lot more explicit