initial Presenter and ViewModel definitions
A basic structure for connecting the `ViewModel` and the `Presenter`
implementations together. A `View` class will implement the `ViewModel`
which in turn will expose the `bind` and `unbind` methods of the `Presenter`.

As the `ViewModel` and the `Presenter` are quite closely bound the `Presenter`
is bound to a generic `ViewModel` class which will expose setter methods in the
implementing `View` class.