reverting to Dagger 1 implementation
Unfortunately it would seem that Dagger2 and Espresso2 are not good friends. Due to issues with running Android tests with Espresso it is necessary to revert back to Dagger1. This situation may change in the future but as it stands now the ability to be able to create and run automated tests out weighs the desire to use Dagger2.

Created a couple of test cases for the `cache` and `mobile` modules to validate tests did in fact run, hence the issues with Dagger2. It seems that the code generation leads to inconsistent results when running application and android-library tests. Tried to work with [android-apt](https://bitbucket.org/hvisser/android-apt) to work around this but to no avail.