provide RemoteEntity module
RemoteEntity module provides simple json remote entity definitions. This is the definition of the remote api model currently being implemented, and allows to simply change the entity definitions if anything changes on the api with necessarily impacting the rest of the application. Also means the remote entity definitions become swappable.