Prototype API to Model factory implementation
Create an Abstract Factory pattern implementation for handling API entity to application model object transformations. `ModelFactory` interface defines a generic implementation and exposes a method taking an API entity and returning an application model object. This method is typed and the implementation of the concrete instance is injected into the using class.

Add test cases to validate the mapping handling for a `SampleEntity` to `Sample` object implementation